Fix Windows compact recovery test crash
