+++ src-tauri/src/agent/mod.rs
@@ -1,4 +1,5 @@
 pub mod channel_routing;
+pub mod compact_recovery;
 pub mod concurrency;
 pub mod config;
 pub mod context;
@@ -9,6 +10,7 @@ pub mod references;
 pub mod session_bus;
 pub mod session_manager;
 pub mod state;
+pub mod tauri_events;
 pub mod tool_approvals;
 pub mod tools;
 pub mod types;
