+++ src-tauri/src/agent/session_manager/compact.rs
@@ -1,4 +1,7 @@
 use super::AgentSessionManager;
+use crate::agent::compact_recovery::{
+    clear_compact_in_flight, handle_compact_error_state, CompactErrorAction,
+};
 use crate::agent::events::AgentEventDispatcher;
 use crate::agent::state::{AgentSession, DeferredWorkflowStep};
 use crate::repositories::{CompactContextRecord, SessionRepository, SessionStatus};
@@ -7,53 +10,6 @@ use std::sync::atomic::Ordering;
 use std::sync::Arc;
 use tokio::sync::RwLock;
 
-async fn clear_compaction_state(
-    active_sessions: &Arc<RwLock<HashMap<String, AgentSession>>>,
-    session_id: &str,
-    clear_last_compacted_tail_id: bool,
-) {
-    let handles = {
-        let active = active_sessions.read().await;
-        active.get(session_id).map(|session| {
-            session.compact_in_flight.store(false, Ordering::SeqCst);
-            session
-                .awaiting_compact_completion
-                .store(false, Ordering::SeqCst);
-            session
-                .finalize_workflow_after_compact
-                .store(false, Ordering::SeqCst);
-            (
-                session.compact_started_at_ms.clone(),
-                clear_last_compacted_tail_id.then(|| session.last_compacted_tail_id.clone()),
-                session.deferred_workflow_step.clone(),
-            )
-        })
-    };
-
-    let Some((
-        compact_started_at_ms_handle,
-        last_compacted_tail_id_handle,
-        deferred_workflow_step_handle,
-    )) = handles
-    else {
-        return;
-    };
-
-    if let Some(last_compacted_tail_id_handle) = last_compacted_tail_id_handle {
-        *last_compacted_tail_id_handle.write().await = None;
-    }
-
-    *deferred_workflow_step_handle.write().await = None;
-    *compact_started_at_ms_handle.write().await = None;
-}
-
-pub async fn clear_compact_in_flight(
-    active_sessions: &Arc<RwLock<HashMap<String, AgentSession>>>,
-    session_id: &str,
-) {
-    clear_compaction_state(active_sessions, session_id, false).await;
-}
-
 pub async fn handle_compact_error_with_dispatcher(
     session_repo: &Arc<dyn SessionRepository>,
     active_sessions: &Arc<RwLock<HashMap<String, AgentSession>>>,
@@ -61,94 +17,17 @@ pub async fn handle_compact_error_with_dispatcher(
     session_id: String,
     error: crate::agent::llm::types::AgentRuntimeError,
 ) -> Result<(), String> {
-    let (
-        was_awaiting,
-        should_finalize_after_compact,
-        deferred_workflow_step,
-        session_name,
-        compact_started_at_ms_handle,
-    ) = {
-        let active = active_sessions.read().await;
-        if let Some(session) = active.get(&session_id) {
-            (
-                session.awaiting_compact_completion.load(Ordering::SeqCst),
-                session
-                    .finalize_workflow_after_compact
-                    .load(Ordering::SeqCst),
-                session.deferred_workflow_step.read().await.clone(),
-                session
-                    .metadata
-                    .name
-                    .clone()
-                    .unwrap_or_else(|| session_id.chars().take(8).collect::<String>()),
-                Some(session.compact_started_at_ms.clone()),
-            )
-        } else {
-            (
-                false,
-                false,
-                None,
-                session_id.chars().take(8).collect::<String>(),
-                None,
-            )
-        }
-    };
-    let elapsed_ms = if let Some(compact_started_at_ms_handle) = compact_started_at_ms_handle {
-        compact_started_at_ms_handle
-            .read()
-            .await
-            .map(|started_at| chrono::Utc::now().timestamp_millis() - started_at)
-    } else {
-        None
-    };
-
-    let error_code = error
-        .details
-        .as_ref()
-        .and_then(|details| details.error_code.as_deref())
-        .unwrap_or("none");
-
-    clear_compaction_state(active_sessions, &session_id, true).await;
-
-    let state_event = crate::agent::llm::types::CompactStateEvent {
-        session_id: session_id.clone(),
-        session_name: Some(session_name),
-        compacting: false,
-        phase: crate::agent::llm::types::CompactStatePhase::Failed,
-        error: Some(error.display_message.clone()),
-    };
-
-    dispatcher.emit_compact_state(state_event)?;
-
-    log::warn!(
-        "❌ Compaction failed: session={}, mode={}, elapsed_ms={}, error_code={}, message={}",
-        session_id,
-        if was_awaiting {
-            "preflight"
-        } else {
-            "background"
-        },
-        elapsed_ms
-            .map(|value| value.to_string())
-            .unwrap_or_else(|| "unknown".to_string()),
-        error_code,
-        error.display_message
-    );
-
-    if was_awaiting || deferred_workflow_step.is_some() {
-        log::warn!(
-            "Blocking compaction failed for session {}. Failing workflow.",
-            session_id
-        );
-        crate::agent::llm::finalize_workflow_error_with_dispatcher(
+    if matches!(
+        handle_compact_error_state(
             session_repo,
             active_sessions,
             dispatcher,
-            session_id,
-            error,
+            session_id.clone(),
+            error
         )
-        .await?;
-    } else if should_finalize_after_compact {
+        .await?,
+        CompactErrorAction::FinalizeWorkflow
+    ) {
         if let Some(app_handle) = crate::state::get_app_handle() {
             crate::agent::lifecycle::update_session_status(
                 session_repo,
@@ -347,7 +226,7 @@ pub async fn handle_compact_response(
                     session_id: session_id.to_string(),
                     reason,
                 };
-                crate::agent::events::emit_agent_event(&manager.app_handle, event)
+                crate::agent::tauri_events::emit_agent_event(&manager.app_handle, event)
                     .map_err(|e| format!("Failed to emit event: {}", e))?;
             }
         }
@@ -409,7 +288,7 @@ pub async fn handle_compact_response(
             session_id: session_id.to_string(),
             reason: crate::agent::events::WorkflowCompletionReason::Natural,
         };
-        crate::agent::events::emit_agent_event(&manager.app_handle, event)
+        crate::agent::tauri_events::emit_agent_event(&manager.app_handle, event)
             .map_err(|e| format!("Failed to emit event: {}", e))?;
     }
