+++ src-tauri/src/mcp/builtin/mcp_manager/operations.rs
@@ -1,4 +1,3 @@
-use crate::agent::events;
 use crate::mcp::builtin::error_guidance::{
     guided_error, missing_param_error, not_found_error, ErrorCategory, SuccessHint, ToolGroup,
 };
@@ -166,7 +165,7 @@ pub async fn register_server(server: &MCPManagerServer, args: Value) -> Result<M
     server.invalidate_cache().await;
 
     // Emit resource updated event for frontend cache revalidation
-    events::emit_resource_updated("mcpServer", "create", Some(name.to_string()));
+    crate::agent::tauri_events::emit_resource_updated("mcpServer", "create", Some(name.to_string()));
 
     let hint = SuccessHint::new(
         format!(
@@ -223,7 +222,7 @@ pub async fn delete_server(server: &MCPManagerServer, args: Value) -> Result<MCP
     server.invalidate_cache().await;
 
     // Emit resource updated event for frontend cache revalidation
-    events::emit_resource_updated("mcpServer", "delete", Some(name.to_string()));
+    crate::agent::tauri_events::emit_resource_updated("mcpServer", "delete", Some(name.to_string()));
 
     let hint = SuccessHint::new(
         format!("Excluded server '{}' from configuration", name),
@@ -323,7 +322,7 @@ pub async fn update_server(server: &MCPManagerServer, args: Value) -> Result<MCP
     server.invalidate_cache().await;
 
     // Emit resource updated event for frontend cache revalidation
-    events::emit_resource_updated("mcpServer", "update", Some(name.to_string()));
+    crate::agent::tauri_events::emit_resource_updated("mcpServer", "update", Some(name.to_string()));
 
     let hint = SuccessHint::new(
         format!("✓ Server configuration updated for '{}' (ID: {})", name, id),
@@ -372,7 +371,7 @@ pub async fn verify_server(server: &MCPManagerServer, args: Value) -> Result<MCP
     server.invalidate_cache().await;
 
     // Emit resource updated event for frontend cache revalidation
-    events::emit_resource_updated("mcpServer", "verify", Some(name.to_string()));
+    crate::agent::tauri_events::emit_resource_updated("mcpServer", "verify", Some(name.to_string()));
 
     match verification_result {
         Ok((tool_count, tools_json)) => {
